clear time labeling implemented
